% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{early_voting_2017}
\alias{early_voting_2017}
\title{NEC 제19대 대통령선거 구시군별 사전투표율
250개 구시군별 전체(0)}
\format{
250개 관측점과 9개 변수를 갖는 데이터프레임
\describe{
\item{결과순서}{문자형, 번호}
\item{선거_id}{문자형, 선거코드(선거일)}
\item{사전투표구분}{문자, 사전투표구분: 전체 0, 첫날 1, 둘째날 2}
\item{시도명}{문자, 시도명}
\item{구시군명}{문자, 구시군명}
\item{선거인수}{문자, 선거인수}
\item{사전투표자수}{문자, 사전투표자수}
\item{사전투표율}{문자, 사전투표율}
\item{순서}{문자, 순서}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
early_voting_2017
}
\description{
18개 시도, 250개 구시군별, 3,510개 읍면동별,
14,464개 투표소별 유권자 남여, 재외국민, 남녀, 거소투표
}
\keyword{datasets}
