% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_code}
\alias{election_code}
\title{중앙선거관리위원회 코드정보}
\format{
79개 관측점과 5개 변수를 갖는 데이터프레임
\describe{
\item{num}{문자형, 번호}
\item{sg_id}{문자형, 선거코드(선거일)}
\item{sg_name}{문자, 선거명}
\item{sg_typecode}{문자, 선거종류코드}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
election_code
}
\description{
선거ID와 선거종류코드, 선거명, 선거구등 선거에 관련된 기초코드
}
\keyword{datasets}
