% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demo_data.R
\name{get_demo_data}
\alias{get_demo_data}
\title{Get demographic data}
\usage{
get_demo_data(table, metric, qmr_con, datamart)
}
\arguments{
\item{table}{Either datamart name or R dataframe}

\item{metric}{Metric to be stratified}

\item{qmr_con}{QMR_DEV connection name}

\item{datamart}{Logical indicating if table argument is a datamart.}
}
\description{
Helper function to add demographic data to input dataset (from either datamart or writing R dataframe to QMR_DEV)
}
