% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gusigun_20220309}
\alias{gusigun_20220309}
\title{선관위 제20대 대통령선거 구시군 코드정보}
\format{
250개 관측점과 5개 변수를 갖는 데이터프레임
\describe{
\item{num}{문자형, 번호}
\item{선거ID}{문자형, 선거코드(선거일)}
\item{구시군명}{문자, 구시군명}
\item{순서}{문자, 순서}
\item{상위시도명}{문자, 시도명}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
gusigun_20220309
}
\description{
선관위 구시군코드 (제20대 대통령선거, 20220309)
OpenAPI활용가이드(코드정보)_v3.6.hwp
}
\keyword{datasets}
