% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gusigun_demo_year_tbl}
\alias{gusigun_demo_year_tbl}
\title{시도, 구시군별 주민등록 인구통계 행정안전부}
\format{
4,520개 관측점과  18개 변수를 갖는 데이터프레임
\describe{
\item{시도명}{문자형, 시도명 전체 명칭}
\item{시도코드}{문자형, 숫자 10자리}
\item{연도}{문자형, 숫자 4자리, 2012 ~ 2020}
\item{행정기관코드}{행정기관코드, 숫자 10 자리코드}
\item{행정기관}{문자형, 행정기관명, 구시군명}
\item{성별}{문자형, 남/여}
\item{x0_9세}{문자형, 0~9세 인구수}
\item{...}{문자형, ...세 인구수}
\item{x100세_이상}{문자형, 100세 이상 인구수}
}
}
\source{
\url{https://jumin.mois.go.kr/}
}
\usage{
gusigun_demo_year_tbl
}
\description{
행정안전부 성별, 연령별 시도, 구시군별 인구현황
2012년 ~ 2021년
}
\keyword{datasets}
