% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demo_plot.R
\name{make_demo_plot}
\alias{make_demo_plot}
\title{Make demographic plots}
\usage{
make_demo_plot(data, metric, demo)
}
\arguments{
\item{data}{Dataframe with demographics}

\item{metric}{Field from dataframe to stratify}

\item{demo}{Demographic fields to plot}
}
\description{
Helper function to make stratified plots by each demographic field.
}
