% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_jumin.R
\docType{data}
\name{sigungu_code}
\alias{sigungu_code}
\title{시군구 마스터 코드}
\format{
A data frame with 246 rows and 4 variables:
\describe{
\item{\code{시도명}}{character 시도명}
\item{\code{시도코드}}{character 10자리 시도, 시군구 코드}
\item{\code{구분}}{character 1: 시도, 2: 시군구}
\item{\code{구시군명}}{character 시군구명}
}
}
\source{
\url{https://jumin.mois.go.kr/}
}
\usage{
sigungu_code
}
\description{
행정안전부 주민등록 인구통계 시군구 마스터 코드
}
\details{
DETAILS
}
\keyword{datasets}
