% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voters_2022_gusigun_tbl}
\alias{voters_2022_gusigun_tbl}
\title{NEC 제20대 대통령선거 유권자
250개 구시군별 유권자 현황}
\format{
250개 관측점과 24개 변수를 갖는 데이터프레임
\describe{
\item{결과순서}{문자형, 번호}
\item{선거_id}{문자형, 선거코드(선거일)}
\item{시도명}{문자, 시도명}
\item{구시군명}{문자, 구시군명}
\item{...}{문자, ...}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
voters_2022_gusigun_tbl
}
\description{
18개 시도 250개 구시군별 유권자 남여, 재외국민, 남녀, 거소투표
}
\keyword{datasets}
