% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voters_2022_sido_tbl}
\alias{voters_2022_sido_tbl}
\title{NEC 제20대 대통령선거 유권자
18개 시도별}
\format{
18개 관측점과 24개 변수를 갖는 데이터프레임
\describe{
\item{결과순서}{문자형, 번호}
\item{선거_id}{문자형, 선거코드(선거일)}
\item{시도명}{문자, 시도명}
\item{...}{문자, ...}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
voters_2022_sido_tbl
}
\description{
선거ID와 선거종류코드, 선거명, 선거구등 선거에 관련된 기초코드
}
\keyword{datasets}
