% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demo_plot.R
\name{make_demo_plot}
\alias{make_demo_plot}
\title{Make demographic plots}
\usage{
make_demo_plot(demo)
}
\arguments{
\item{demo}{Dataframe with demographic fields to plot}
}
\description{
Helper function to make stratified plots by each demographic field.
}
