% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_demos.R
\name{screen_demos}
\alias{screen_demos}
\title{Stratify dataset by demographic variables}
\usage{
screen_demos(table, metric, qmr_con, datamart = TRUE)
}
\arguments{
\item{table}{Either datamart name or R dataframe}

\item{metric}{Metric to be stratified}

\item{qmr_con}{QMR_DEV connection name}

\item{datamart}{Logical indicating if table argument is a datamart. Defaults to TRUE.}
}
\description{
Function to stratify a metric from an input dataset by race/ethnicity, primary language, and payer type.
}
